% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRound.R
\name{extractRound}
\alias{extractRound}
\title{Scrape future games from a round}
\usage{
extractRound(round)
}
\arguments{
\item{round}{Scalar indicating the roound code}
}
\description{
Scrape future games from a round
}
