% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit point-spread model}
\usage{
fitModel(y, x, init)
}
\arguments{
\item{y}{Numeric vector of point spreads.}

\item{x}{Numeric vector of Elo rating differences.}

\item{init}{Named vector with starting values for the parameters
to be optimized.}
}
\description{
Fit point-spread model
}
