% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElo.R
\name{getElo}
\alias{getElo}
\title{Get Elo ratings for several seasons}
\usage{
getElo(df, k = 25, home_adv = 100, s = 400, carry = 0.8)
}
\arguments{
\item{carry}{}
}
\description{
Get Elo ratings for several seasons
}
