% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getExpectedProb}
\alias{getExpectedProb}
\title{Compute expected win probability}
\usage{
getExpectedProb(r_team, r_opp, home_adv, s)
}
\arguments{
\item{s}{}
}
\description{
Compute expected win probability
}
