% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatestElo.R
\name{getLatestElo}
\alias{getLatestElo}
\title{Get latest Elo ratings}
\usage{
getLatestElo(df, .season, as.list = TRUE)
}
\arguments{
\item{df}{Data frame of Elo ratings as returned by \code{getElo()}
in long format.}
}
\description{
\code{getLatestElo} returns the latest Elo ratings for all teams in the given season
}
