% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getMovMultiplier}
\alias{getMovMultiplier}
\title{Compute margin of victory multiplier}
\usage{
getMovMultiplier(points_diff, elo_diff)
}
\arguments{
\item{elo_diff}{}
}
\description{
Compute margin of victory multiplier
}
