% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{getParTable}
\alias{getParTable}
\title{Compute mean, sd, and credible interval of model parameters}
\usage{
getParTable(fit)
}
\arguments{
\item{fit}{List with model fit}
}
\description{
Compute mean, sd, and credible interval of model parameters
}
