% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonElo.R
\name{getSeasonElo}
\alias{getSeasonElo}
\title{Compute Elo ratings for a single season}
\usage{
getSeasonElo(df, k = 25, home_adv = 100, s = 400, initial_elo)
}
\arguments{
\item{df}{Data frame of results (as gmresults)}

\item{k}{Scalar specifying the k factor}

\item{home_adv}{Scalar specifying home court advantage (in Elo points)}

\item{s}{Scalar specifying the scale parameter}

\item{initial_elo}{Named list with each element specifying a team Elo rating}
}
\description{
Compute Elo ratings for a single season
}
