% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSeries.R
\name{getSeriesWinner}
\alias{getSeriesWinner}
\title{Get winner from a Playoff series}
\usage{
getSeriesWinner(df)
}
\arguments{
\item{df}{Data frame with Playoff series simulation results}
}
\description{
Get winner from a Playoff series
}
