% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getTiedTeams}
\alias{getTiedTeams}
\title{Get teams with same number of wins}
\usage{
getTiedTeams(df)
}
\arguments{
\item{df}{Data frame of Elo ratings in long format}
}
\description{
Get teams with same number of wins
}
