% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getWinsPast}
\alias{getWinsPast}
\title{Get table of wins and overall point differential from past games}
\usage{
getWinsPast(df)
}
\arguments{
\item{df}{Data frame with past Elo ratings in long format}
}
\description{
Get table of wins and overall point differential from past games
}
