% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getWinsSim}
\alias{getWinsSim}
\title{Get table of wins and overall point differential from simulated games}
\usage{
getWinsSim(df)
}
\arguments{
\item{df}{Data frame of simulated games as returned by \code{simElo()}}
}
\description{
Get table of wins and overall point differential from simulated games
}
