% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{mergeWins}
\alias{mergeWins}
\title{Merge past and simulated win tables}
\usage{
mergeWins(past, sim)
}
\arguments{
\item{past}{Data frame with past win table}

\item{sim}{Data frame with simulated win table}
}
\description{
Merge past and simulated win tables
}
