% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotEloLong.R
\name{pivotEloLong}
\alias{pivotEloLong}
\title{Convert data frame of Elo ratings from wide to long}
\usage{
pivotEloLong(df)
}
\arguments{
\item{df}{Data frame of Elo ratings as returned by \code{getElo()}}
}
\description{
Convert data frame of Elo ratings from wide to long
}
