% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{predictPointDiff}
\alias{predictPointDiff}
\title{Predict a single mu observation for a single Elo difference}
\usage{
predictPointDiff(fit, x)
}
\arguments{
\item{fit}{List with model fit}

\item{x}{Scalar of Elo difference in a game}
}
\description{
Predict a single mu observation for a single Elo difference
}
