% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFinal4.R
\name{simFinal4}
\alias{simFinal4}
\title{Simulate Final Four results}
\usage{
simFinal4(teams, elo, fit)
}
\arguments{
\item{teams}{Character vector of length 4 with the name of the teams
that won the Playoffs. Must be in the same order of Playoffs matchups.}

\item{elo}{Numeric vector of length 4 with the Elo ratings of the teams
that won the Playoffs. Must be in the same order of Playoffs matchups.}

\item{fit}{List with the point-spread model fit}
}
\description{
Simulate Final Four results
}
