% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simProbability.R
\name{simProbability}
\alias{simProbability}
\title{Get simulation probabilities}
\usage{
simProbability(n.sim, fgames, pgames, fit)
}
\arguments{
\item{n.sim}{Scalar indicating the number of simulations to perform}

\item{fgames}{Data frame of future games in wide format (as in rschedule)}

\item{pgames}{Data frame of past games in long format (as in elohist)}

\item{fit}{List of fitted point-spread model}
}
\description{
Get simulation probabilities
}
