% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRegSeason.R
\name{simRegSeason}
\alias{simRegSeason}
\title{Simulate regular season}
\usage{
simRegSeason(fgames, pgames, fit)
}
\arguments{
\item{fgames}{Data frame with future games in wide format (as in rschedule)}

\item{pgames}{Data frame with past games in long format (as in elohist)}
}
\description{
Simulate regular season
}
