% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simElo.R
\name{simElo}
\alias{simElo}
\title{Simulate games using Elo ratings}
\usage{
simElo(df, k, home_adv, s, initial_elo)
}
\arguments{
\item{df}{Data frame with all the games (as in rschedule)}

\item{k}{Scalar indicating the k factor}

\item{home_adv}{Scalar indicating home court advantage (in Elo points)}

\item{s}{Scalar indicating the scale parameter}

\item{initial_elo}{Named list with each element specifying a team Elo rating}
}
\description{
Simulate games using Elo ratings
}
